services/coletor-python/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
kafka-python==2.0.2
requests==2.31.0